pyproject.toml
[tool.poetry]
name = "jfd"
version = "0.1.0"
description = ""
authors = ["Beknazar <abdyldaevgroup@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
pillow = "^9.5.0"
drf-yasg = "^1.21.6"
celery = "^5.3.1"
redis = "^4.6.0"
pyjwt = "^2.7.0"
djangorestframework-simplejwt = "^5.2.2"
gunicorn = "^20.1.0"
gevent = "^22.10.2"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
